don't store header in tag
When parsing a whole tag, the header we read is mostly useless to the
user. The version will always be 2.4, because we upgrade the tag, and
the size will not reflect changes to frames. Only the flags are of
interest, which we can store directly in the tag.

If a user is interested in the actual header, he can call ParseHeader
instead.